Update file type
For some reason, the type isn't getting "stamped" onto the files
appropriately; they show up with a dynamic UTI, and the document controller
refuses to open them.

I'm this close to just hacking around it by checking file extension
and trying to open using my Document type directly, except that I'd like
to be able to Cmd-o in Finder to open a slicedImage file.