Share schemes for QuickLook plugin
Building the invocation qlmanage was a bit of a bear,
but at last, I'm in a place to begin debugging the actual running plugin.

Needed to specify the specific generator to use to prevent it from
preferring an older generator embedded in the app bundle.

The .slicedimage document path will need to be changed by anyone using
this scheme; I'm sure qlmanage will kvetch loudly enough to make that clear.